+++ octopose/local_deploy.py
@@ -41,9 +41,10 @@ class LocalDeploy:
                 args = "powershell.exe {0}".format(step_path)
             else:
                 args = "c:\\windows\\sysnative\\cmd.exe /c powershell.exe {0}".format(step_path)
-            self.subprocess_runner.run(args, "Running of {0} failed".format(step_path))
+            return self.subprocess_runner.run(args, "Running of {0} failed".format(step_path))
         else:
             print("Can't find path - skipping this file")
+            return True
 
     def deploy(self, data):
         """deploy_local will use the manifest to deploy all packages to the local machine"""
@@ -52,7 +53,9 @@ class LocalDeploy:
             shutil.rmtree(staging)
         os.makedirs(staging, mode=0o777)
 
+        deployment_results = []
         for key, value in data['Projects'].items():
+            successful_deployment = True
             project_name = key
             proj_id = octo.get_project_id(project_name)
             if 'Version' not in value:
@@ -68,9 +71,17 @@ class LocalDeploy:
                 print("- NuGet - {0} - {1}".format(package_name, version))
                 self.nu.get_deployable(package_name, version, staging)
 
-                self.invoke_deploy("{0}\{1}.{2}\PreDeploy.ps1".format(staging, package_name, version))
-                self.invoke_deploy("{0}\{1}.{2}\Deploy.ps1".format(staging, package_name, version))
-                self.invoke_deploy("{0}\{1}.{2}\PostDeploy.ps1".format(staging, package_name, version))
+                for script in ["{0}\{1}.{2}\PreDeploy.ps1", "{0}\{1}.{2}\Deploy.ps1", "{0}\{1}.{2}\PostDeploy.ps1"]:
+                    if not self.invoke_deploy(script.format(staging, package_name, version)):
+                        successful_deployment = False
+                        break
+
+                if not successful_deployment:
+                    break
+
+            deployment_results.append((project_name, release_version, successful_deployment))
+
+        print_deployment_results(deployment_results)
 
 
 def get_package_versions(proj_id, release_version, package_ids_to_deploy):
@@ -95,3 +106,13 @@ def get_package_versions(proj_id, release_version, package_ids_to_deploy):
 
 def is_64_bit_python_installation():
     return sys.maxsize > 2**32
+
+
+def print_deployment_results(deployment_results):
+    print("-- Deployment Results --")
+    for result in deployment_results:
+        if result[2]:
+            status = "Success"
+        else:
+            status = "Failure"
+        print("{0} - {1} - {2}".format(result[0], result[1], status))